pyproject.toml
[project]
name = "vacation-rentals-agent"
version = "0.1.0"
description = "A vacation rental customer service agent for policy questions and guidance"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk[a2a]>=1.18.0",
    "litellm>=1.65.0",
    "httpx>=0.28.0",
    "uvicorn>=0.34.0",
    "loguru>=0.7.3",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
